replace animation public files with a submodule
I bet that submodules are a terrible idea, but no better way to learn
why than to try them myself. I've put the animations in public so that
each page can stand alone and the corresponding css assets can be
served. Theorietically if I do it right I'll just need to merge in new
changes to the submodule to update the blog.