create page that links to the css3_animations submodle
this approach is pretty horrible, but at least it's a first pass. I
don't like the idea of tailoring the directory structure of the
animations repo in order to fit with the blog, but in order to serve the
css assets correctly I need to update the path in the <link>

by reading in the html file and searching for + updating the path, then
serving the updated html, this gets the job done but my god is it ugly
and vulnerable.