just point to the static files
all of the hoops I jumped through revolved around the idea of serving
the animations through a sinatra path when they were all contained in
the public directory. I'd like for all of these files to not sit in the
public directory and to use a cleaner route structue but there isn't a
reason for them not to be there.