.scss-ify style.css
it's not a super thorough refactoring but this pulls all of the main
colors that I use into variables so I can easily adjust them.

I'm still undecided on which files to check in. I think once I've
converted all of the current .css files to .scss then I'll just add all
of public/stylesheets/*.css to .gitignore?