create section nav
this site was just supposed to be a blog before, so there was only one
section of pages to deal with. now that there are more there needs to be
a way to navigate between them.

I went with a hover-over dropdown menu next to the subtitle. I haven't
teseted to accessability of this yet but I bet it's not very good.