ajax load in the animations!
each animation requires two parts: the #animations div and the
corresponding stylesheet. since I have the urls listed and the css files
have the same name I can just load in these parts with javascript.

in order to insert the javascript and styles accordingly I had to insert
a couple of `yield_content` sections in the layout. I need to take
advantage of these on other pages now.