bust out the tricks to make the animation responsive
this one was tough, as the end result isn't perfect, but I'm happy with
it for now. the animations box will shrink as needed to fit within the
viewport. since the animations themselves aren't larger than 180x180
they should fit on the smallest of screens. however, at really low
resolutions the animations start to become off-center within the
container.

I'm not supe thrilled about the animation names being /below/ the
animation container when the screen isn't wide enough to support them
side-by-side, but the container has the instructions and could be missed
entirely if it's below.