flatten sections nav
the old sections nav was a cute little drop down that appeared when you
hovered over it, but it wasn't keyboard accessible at all. after trying
a few tricks I couldn't really get it to work, so I've decided to just
flatten it. this is better anyway because it makes the different
sections easily viewable.

resolves issue #7