reorganize and test latest_posts
I've moved latest_posts into the SiteHelpers module since it's small
enough that I don't feel it's cluttered and it makes it easier to test.
the test itself is rather complicated but it ensures that a random set
of posts are created that don't match my specific criteria, so it should
never give a flase positive result.