blowing off the dust
it's been a while since I've updated anything and a lot of stuff has
been upgraded. One thing I try to do with personal projects is to keep
them up-to-date so everything gets a bump!

I'm also using rbenv now just to see what's it like so that's the flavor
of the day.