nepeta cataria
it's been over a month since we made this game of LD34 but now I'm
finally hosting it! I ended up shrinking the menu screen down because it
wasn't auto-scaling to a lower resolution. that's one part of unity that
I don't really understand but I was able to make a custom build for my
site.