use print css styles to format resume for pdf
rather than running a markdown to pdf converter, I can utilize CSS to
make the pdf export of my resume look a lot nicer. this introduces a
specific set of print styles to the hire page to hide the header and
ensure define text spacing in terms of physical measurements for the
resume text.